# lib/api-zod/src/generated/types/transactionStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PayRald API — consumer financial layer of the RALD ecosystem
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionStatsMonthlyVolumeItem } from './transactionStatsMonthlyVolumeItem';

export interface TransactionStats {
  totalSent: number;
  totalReceived: number;
  totalFees: number;
  feesSaved: number;
  transferCount: number;
  paymentCount: number;
  withdrawalCount: number;
  monthlyVolume: TransactionStatsMonthlyVolumeItem[];
}