# lib/api-zod/src/generated/types/transferPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PayRald API — consumer financial layer of the RALD ecosystem
 * OpenAPI spec version: 0.1.0
 */
import type { Transfer } from './transfer';

export interface TransferPage {
  data: Transfer[];
  /** @nullable */
  nextCursor?: string | null;
  total: number;
}