# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PayRald API — consumer financial layer of the RALD ecosystem
 * OpenAPI spec version: 0.1.0
 */
import type { UserActivatedTypesItem } from './userActivatedTypesItem';

export interface User {
  id: string;
  /** @handle */
  raldId: string;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  activatedTypes: UserActivatedTypesItem[];
  /**
     * @minimum 0
     * @maximum 3
     */
  kycTier: number;
  trustScore: number;
  createdAt: Date;
}